+++ eldonationtracker/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "6.0.2"
+__version__ = "6.1.0"
 
 base_api_url: str = "https://www.extra-life.org/api"